+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next Release
 
 ### Features Added
+- Introduced `--slashing-protection-db-pool-configuration-file` to specify Hikari connection pool configuration file.
 - Upgraded gradle and various plugin versions. Switched to new dependency license reporting plugin. Project can now be compiled against JDK 16.
 - Introduced --network cli option for Eth2 mode. Defaults to mainnet. Should match the option used by Teku at runtime.
 - Upgraded Teku libraries.